Restructure ui components
