+++ gui/src/components/maps/AladinInteractions.jsx
