+++ gui/src/components/maps/ErrorOverlay.jsx
