+++ gui/src/components/maps/HeaderControls.jsx
