+++ gui/src/components/maps/StatusBar.jsx
