+++ gui/src/components/pipeline/PipelineAppSelector.jsx
