+++ gui/src/components/pipeline/PipelineProgress.jsx
