+++ gui/src/components/ui/ImageModal.jsx
